+++ src/contracts/fungible/data/outcoins.rs
@@ -31,7 +31,7 @@ use crate::error::ParseError;
 #[display_from(Debug)]
 pub struct Outcoins {
     pub coins: f32,
-    pub vout: u16,
+    pub vout: u32,
     pub txid: Option<Txid>,
 }
 
@@ -75,7 +75,7 @@ impl StrictDecode for Outcoins {
     fn strict_decode<D: io::Read>(mut d: D) -> Result<Self, Self::Error> {
         Ok(Self {
             coins: f32::strict_decode(&mut d)?,
-            vout: u16::strict_decode(&mut d)?,
+            vout: u32::strict_decode(&mut d)?,
             txid: Option::<Txid>::strict_decode(&mut d)?,
         })
     }
